Move datagen to neoforge subpackage
