+++ src/main/java/pokefenn/totemic/neoforge/TotemicNeoMod.java
@@ -13,16 +13,6 @@ import pokefenn.totemic.api.TotemicAPI;
 import pokefenn.totemic.apiimpl.registry.RegistryApiImpl;
 import pokefenn.totemic.block.totem.entity.TotemBaseBlockEntity;
 import pokefenn.totemic.compat.kubejs.TotemicKubeEventHandler;
-import pokefenn.totemic.data.TotemicAdvancementProvider;
-import pokefenn.totemic.data.TotemicBlockStateProvider;
-import pokefenn.totemic.data.TotemicBlockTagsProvider;
-import pokefenn.totemic.data.TotemicDamageTypeTagsProvider;
-import pokefenn.totemic.data.TotemicDataMapProvider;
-import pokefenn.totemic.data.TotemicDatapackEntryProvider;
-import pokefenn.totemic.data.TotemicEntityTypeTagsProvider;
-import pokefenn.totemic.data.TotemicItemTagsProvider;
-import pokefenn.totemic.data.TotemicLootTableProvider;
-import pokefenn.totemic.data.TotemicRecipeProvider;
 import pokefenn.totemic.init.ModBlockEntities;
 import pokefenn.totemic.init.ModBlocks;
 import pokefenn.totemic.init.ModContent;
@@ -32,6 +22,16 @@ import pokefenn.totemic.init.ModEntityTypes;
 import pokefenn.totemic.init.ModItems;
 import pokefenn.totemic.init.ModMobEffects;
 import pokefenn.totemic.init.ModSounds;
+import pokefenn.totemic.neoforge.datagen.TotemicAdvancementProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicBlockStateProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicBlockTagsProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicDamageTypeTagsProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicDataMapProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicDatapackEntryProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicEntityTypeTagsProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicItemTagsProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicLootTableProvider;
+import pokefenn.totemic.neoforge.datagen.TotemicRecipeProvider;
 import pokefenn.totemic.neoforge.handler.PlayerInteract;
 import pokefenn.totemic.network.NetworkHandler;
